cql3: expr: Add as_column_maybe_subscripted
Add a convenience function that allows to convert
a reference to expression to column_maybe_subscripted.

It will be useful in a moment.

For now part of it must be commented out
because subscript is not in the expression variant yet.

It will be uncommented once subscript is finally added
to expression.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>