cql3: expr: Handle subscript in extract_clustering_prefix_restrictions
extract_clustering_prefix_restrictions collects restrictions
on clustering key columns.

In case we encounter col[sub] we treat it as a restriction on col
and add it to the result.

This seems to make some sense and is in line with the current behaviour
which doesn't check whether a column is subscripted at all.

The code has been copied from column_value& handler.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>