test: flat_mutation_reader_test: harden test_fragmenting_and_freezing
test_fragmenting_and_freezing generates a random mutation and freezes it,
checking the number of frozen fragments against the mutation itself.

However, once moved to flat_mutation_reader_v2(), this breaks. The issue
is that v2 readers canonicalize the mutation, and so if it has overlapping
range tombstones then the number of fragments will be smaller than the
number that can be expected looking at the non-canonical mutation.

Fix this by passing the mutation through make_flat_mutation_reader_from_mutations().
This generates the same number of fragments now (it does not canonicalize),
but when the test is updated to make_flat_mutation_reader_from_mutations_v2,
both the count and the generated fragments will be canonicalized in the
same way, so the test can continue passing.